Files reorganized efficiently
