+++ frontend/src/Components/Forms/SigninForm/Signin.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Form, Row, Col, Button, Container, Spinner, FloatingLabel } from "react-bootstrap";
+import { Form, Row, Col, Button, Container, FloatingLabel } from "react-bootstrap";
 import { Link } from "react-router-dom";
 
 const SignIn = () => {
